Add files via upload
Predicting Stock Market
prices in today’s era is very valuable
and profitable. It is not an easy task
though. It can be affected by
multiple factors, such as gross
domestic product, politics,
exchange rates, gold prices, etc. In
this paper, we aim to use different
machine-learning models to predict
the stock index of the NASDAQ
Composite to figure out the best
model to do so.